requirements.txt
bleach==2.1.3
botocore==1.12.205
certifi==2017.11.5
chardet==3.0.4
click==6.7
construct==2.5.3
Django==2.0.2
django-appconf==1.0.3
django-cors-headers==3.0.2
django-debug-toolbar==1.9.1
django-haystack==2.8.1
django-shortuuidfield==0.1.3
docutils==0.14
future==0.16.0
haystack==0.42
html5lib==1.0.1
idna==2.6
itsdangerous==0.24
jmespath==0.9.4
lxml==4.2.1
MarkupSafe==1.0
mysqlclient==1.4.2.post1
pefile==2019.4.18
pilkit==2.0
Pillow==6.0.0
PyMySQL==0.9.3
python-dateutil==2.8.0
python-decouple==3.1
python-memcached==1.59
python-ptrace==0.9.3
pytz==2017.3
requests==2.21.0
s3transfer==0.2.1
shortuuid==0.5.0
six==1.11.0
sqlparse==0.2.4
urllib3==1.24.2
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==4.1.3
Whoosh==2.7.4
